Moved authJwt middleware out of server.js in prepartion for user model/router updates.
