+++ api/server.js
@@ -3,10 +3,7 @@ const server = express();
 
 const helmet = require('helmet');
 const cors = require('cors');
-
-const jwt = require('express-jwt');
-const jwtAuthz = require('express-jwt-authz');
-const jwksRsa = require('jwks-rsa');
+const checkJwt = require('../middleware/authJWT.js')
 
 server.use(express.urlencoded({ extended: true }));
 server.use(express.json());
@@ -18,31 +15,8 @@ server.get('/', (req, res) => {
   res.send(`<h1>Server live</h1>`);
 });
 
-// # Auth0 Middleware # //
-const authConfig = {
-  domain: 'key-conservation.auth0.com',
-  audience: 'https://key-conservation'
-};
-
-// Define middleware that validates incoming bearer tokens
-// using JWKS from YOUR_DOMAIN
-const checkJwt = jwt({
-  secret: jwksRsa.expressJwtSecret({
-    cache: true,
-    rateLimit: true,
-    jwksRequestsPerMinute: 5,
-    jwksUri: `https://${authConfig.domain}/.well-known/jwks.json`
-  }),
-
-  audience: authConfig.audience,
-  issuer: `https://${authConfig.domain}/`,
-  algorithm: ['RS256']
-});
-console.log(checkJwt);
-// # End Auth0 Middleware # //
-
 // defined routes
-server.use('/api/users', checkJwt, require('../users/usersRouter'));
+server.use('/api/users', require('../users/usersRouter'));
 server.use('/api/campaigns', checkJwt, require('../campaigns/campRouter'));
 server.use(
   '/api/updates',
@@ -52,4 +26,4 @@ server.use(
 server.use('/api/comments', checkJwt, require('../comments/commentsRouter'));
 server.use('/api/social', checkJwt, require('../social/socialRouter'));
 
-module.exports = server;
+module.exports = server
