+++ auth/authJwt.js
@@ -1,0 +1,28 @@
+const jwt = require('express-jwt');
+const jwtAuthz = require('express-jwt-authz');
+const jwksRsa = require('jwks-rsa');
+
+// # Auth0 Middleware # //
+const authConfig = {
+  domain: 'key-conservation.auth0.com',
+  audience: 'https://key-conservation'
+};
+
+// Define middleware that validates incoming bearer tokens
+// using JWKS from YOUR_DOMAIN
+const checkJwt = jwt({
+  secret: jwksRsa.expressJwtSecret({
+    cache: true,
+    rateLimit: true,
+    jwksRequestsPerMinute: 5,
+    jwksUri: `https://${authConfig.domain}/.well-known/jwks.json`
+  }),
+
+  audience: authConfig.audience,
+  issuer: `https://${authConfig.domain}/`,
+  algorithm: ['RS256']
+});
+console.log(checkJwt);
+
+module.exports = checkJWT
+// # End Auth0 Middleware # //
